requirements.txt

jittor>=1.3.0
numpy>=1.19.0
pillow>=8.0.0
tqdm>=4.60.0
